refector(mentor): component 분리, 에러 메세지 못보내는 부분 수정
