+++ src/App.js
@@ -1,9 +1,10 @@
 import './css/reset.css';
 import './css/root.css';
+import './components/components.css';
 import Header from './components/Header';
-import Shared from './components/shared/Shared';
 import Footer from './components/Footer';
-import Folder from './components/folder/Folder';
+import Shared from './pages/Shared';
+import Folder from './pages/Folder';
 import { useFetch} from './hooks/useFetch';
 import { AccountContext } from './contexts/AccountContext';
 import { Route, Routes } from 'react-router-dom';
@@ -11,9 +12,8 @@ import { Route, Routes } from 'react-router-dom';
 
 function App() {
   const {data: userData, errorMessage} = useFetch("users/1", 1);
-  if(!userData) return;
-
   
+  if(!userData) return;
   return (
     <AccountContext.Provider value={{account: userData.data[0], errorMessage: errorMessage}}>
       <div className="App">
