+++ src/components/AddInputSection.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import AddLink from './AddLink';
+
+
+
+const AddInputSection = () => {
+    
+    return (
+        <div className="section-title section-title-first">
+            <AddLink/>
+        </div>
+    );
+};
+
+export default AddInputSection;
